Fix the Update and Delete paths to allow for partial matching and allow for an empty ReporterInstanceId
